requirement.txt
certifi==2020.4.5.1
numpy==1.18.3
Pillow==6.1.0
pydicom==1.4.2
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
wincertstore==0.2
